# Replace encoding with compression. Bump version