--- a/lib/zjson-bin.js
+++ b/lib/zjson-bin.js
@@ -9,7 +9,6 @@
  * - Log is not efficient to allow
  * - the zjon format should carry some flags (is it compressed and needs decompression)
  * - dictionary index is 32based... not optimum where there is more than 32 field names in the object.
- * - bjson could be implemented in msgpack-lite. other gzip-based libs were not efficient compression as it would not optimize the javascript obj.
  *
  * DEVELOPMENT
  * When enhancing this lib, recommendation
@@ -27,7 +26,7 @@
   function() {
     registerLib(this, factory);
 
-    function factory(_, msgpack, logger) {
+    function factory(_, pako, logger) {
       'strict mode';
 
       const zjsonbin = {
@@ -67,35 +66,15 @@
         }
         const start = Date.now();
 
-        const cJson = pack(obj);
-        const serializedObj = msgpack.encode(cJson);
-        const string = abs2str(serializedObj);
+        const serializedObj = pako.deflate(JSON.stringify(obj), { to: 'string' });
 
         if (zjsonbin.debug === 2) {
-          const compression = Math.round(Number(100 * string.length / objLength) * 10) / 10;
-          logger.debug('Serialize object to size ' + getSizeFormat(string.length) + ', compressed to ' + compression + '% in ' + getTimeFormat(Date.now() - start));
+          const compression = Math.round(Number(100 * serializedObj.length / objLength) * 10) / 10;
+          logger.debug('Serialize object to size ' + getSizeFormat(serializedObj.length) + ', compressed to ' + compression + '% in ' + getTimeFormat(Date.now() - start));
         } else if (zjsonbin.debug) {
-          logger.debug('Serialize object to size ' + getSizeFormat(string.length) + ' in ' + getTimeFormat(Date.now() - start));
+          logger.debug('Serialize object to size ' + getSizeFormat(serializedObj.length) + ' in ' + getTimeFormat(Date.now() - start));
         }
-        return string;
-
-        // const jsonText = JSON.stringify(obj);
-        // const jsonT = Date.now();
-
-        // const jsonParse = jsonify(obj);
-        // const jsonP = Date.now();
-
-        // const packed = msgpack.encode(jsonParse);
-        // //jsonpack.pack(jsonParse);
-        // const packedT = Date.now();
-        // const compression = Math.round(Number(100 * packed.length / jsonText.length) * 10) / 10;
-
-        // debug && logger.debug('Serialize object to size %b, compressed to %b - jtext %b - jparse %b - packed %b',
-        //     getSizeFormat(packed.length), compression + '%',
-        //     getTimeFormat(jsonT - start),
-        //     getTimeFormat(jsonP - jsonT),
-        //     getTimeFormat(packedT - jsonP));
-        // return packed;
+        return serializedObj;
       }
 
       /**
@@ -104,17 +83,17 @@
        * @param {String} stringData
        * @return {Object} obj
        */
-      function deserialize(stringData) {
-        if (zjsonbin.disabled || _.isEmpty(stringData)) {
+      function deserialize(binaryString) {
+        if (zjsonbin.disabled || _.isEmpty(binaryString)) {
           return null;
         }
 
         const start = Date.now();
-        const serializedObj = str2ab(stringData);
-        const compactJson = msgpack.decode(serializedObj);
-        const obj = unpack(compactJson);
+        const compactJson = pako.inflate(binaryString, { to: 'string' });
+
+        const obj = JSON.parse(compactJson);
         zjsonbin.debug && logger.debug('Deserialized object of size ' +
-          getSizeFormat(stringData.length) + ' in ' + getTimeFormat(Date.now() - start));
+        getSizeFormat(binaryString.length) + ' in ' + getTimeFormat(Date.now() - start));
 
         return obj;
       }
@@ -125,138 +104,6 @@
 
       function getTimeFormat(lap) {
         return lap >= 1000 ? Math.round((lap) / 100) / 10 + 's' : lap + 'ms';
-      }
-
-      /**
-       * take an object,
-       * jsonify it and factor its property names (dictionary use) to reduce its size
-       * @param {*} obj
-       * @returns {Object} optimized json object
-       */
-      function pack(obj) {
-        const dictionaryNames = {},
-          dictionary = [];
-        let n = 0;
-        const r = _.cloneDeepWith(obj, processProperty);
-        return [dictionary, r];
-
-        function processProperty(objProperty) {
-          if (objProperty && objProperty.toJSON) {
-            objProperty = objProperty.toJSON();
-          }
-          if (_.isArray(objProperty)) {
-            return _.map(objProperty, function(item) {
-              if (_.isNil(objProperty)) {
-                return null;
-              }
-              return _.cloneDeepWith(item, processProperty);
-            });
-          }
-          if (_.isObject(objProperty)) {
-            // Data coming in can use new Number(42) instead of 42
-            if (_.isNumber(objProperty)) {
-              return Number(objProperty);
-            }
-            // Data coming in can use new String('test') instead of 'test'
-            if (_.isString(objProperty)) {
-              return String(objProperty);
-            }
-
-            let property, d;
-            const newObj = {};
-            for (property in objProperty) {
-              if (!_.isNil(objProperty[property]) && !_.isFunction(objProperty[property])) {
-                d = dictionaryNames[property];
-                if (_.isNil(d)) {
-                  d = (n++).toString(36);
-                  dictionaryNames[property] = d;
-                  dictionary.push(property);
-                }
-
-                newObj[d] = _.cloneDeepWith(objProperty[property], processProperty);
-              }
-            }
-            return newObj;
-          }
-          return objProperty;
-        }
-      }
-
-      /**
-       * take a packed(optimized) json object and unpack to recover its original form.
-       *
-       * @param {Object} packedObj
-       */
-      function unpack(packedObj) {
-        const dictionary = packedObj[0];
-        const r = _.cloneDeepWith(packedObj[1], processProperty);
-        return r;
-
-        function processProperty(objProperty) {
-          if (_.isArray(objProperty)) {
-            return _.map(objProperty, function(item) {
-              if (_.isNil(objProperty)) {
-                return null;
-              }
-              return _.cloneDeepWith(item, processProperty);
-            });
-          }
-          if (_.isObject(objProperty)) {
-            // console.log(objProperty.constructor.name);
-            const newObj = {};
-            for (let property in objProperty) {
-              if (!_.isNil(objProperty[property]) && !_.isFunction(objProperty[property])) {
-                let d = dictionary[parseInt(property, 36)];
-                newObj[d] = _.cloneDeepWith(objProperty[property], processProperty);
-              }
-            }
-            return newObj;
-          }
-          return objProperty;
-        }
-      }
-
-      /**
-       * Convert a string to a buffer
-       *
-       * note that this is only correct if your buffer only contains non-multibyte ASCII characters:
-       *
-       * @param {String} str
-       *
-       * @returns {Uint8Array}
-       */
-      function str2ab(str) {
-        let buf = new ArrayBuffer(str.length); // 2 bytes for each char
-        let bufView = new Uint8Array(buf);
-        for (let i = 0, strLen = str.length; i < strLen; i++) {
-          bufView[i] = str.charCodeAt(i);
-        }
-        return bufView;
-      }
-
-      /**
-       * Convert a buffer to a string
-       *
-       * and prevent Maximum call stack size exceeded
-       *
-       * https://stackoverflow.com/questions/12710001/how-to-convert-uint8-array-to-base64-encoded-string/12713326#12713326
-       *
-       * note that this is only correct if your buffer only contains non-multibyte ASCII characters:
-       * @param {Uint8Array} u8a
-       *
-       * @return {String}
-       */
-      function abs2str(u8a) {
-        let CHUNK_SZ = 0x8000;
-        if (u8a.length < CHUNK_SZ) {
-          return String.fromCharCode.apply(null, u8a);
-        }
-
-        let c = [];
-        for (let i = 0; i < u8a.length; i += CHUNK_SZ) {
-          c.push(String.fromCharCode.apply(null, u8a.subarray(i, i + CHUNK_SZ)));
-        }
-        return c.join("");
       }
 
       /**
@@ -482,17 +329,17 @@
         // register for node
         logger = require('zimit-zlog').getLogger('zjsonbin');
         logger.setLevel('all');
-        factory(require("lodash"), require("msgpack-lite"), logger);
+        factory(require("lodash"), require("pako"), logger);
       } else if (typeof define === "function" && define.amd) {
         // not tested.
         logger = { debug: console.debug };
-        define(["_", "msgpack-lite"], logger, function(lodash, msgpack) {
-          factory(lodash, msgpack, logger);
+        define(["_", "pako"], logger, function(lodash, pako) {
+          factory(lodash, pako, logger);
         });
       } else {
         // register lib for browser if dependency are loaded.
         logger = { debug: console.debug };
-        factory(global._, global.msgpack, logger);
+        factory(global._, global.pako, logger);
       }
     }
   }