--- a/package.json
+++ b/package.json
@@ -1,6 +1,6 @@
 {
   "name": "zjsonbin",
-  "version": "1.0.1",
+  "version": "2.0.0",
   "description": "Json serialization/deserialization Tool Set",
   "main": "lib/zjson-bin.js",
   "keywords": [
@@ -21,7 +21,7 @@
   "license": "MIT",
   "dependencies": {
     "lodash": "^4.0.0",
-    "msgpack-lite": "^0.1.26",
+    "pako": "^1.0.11",
     "promise": "^7.1.1",
     "zimit-zlog": "1.0.3"
   },