Added compression 
Reduce overhead with transforming To/From ArrayBuffer
Reduced packing overhead that used multiple deep clones
Default zlib compression is 6

Performance metrics:
Before:
```
2020-03-20 06:20:02 EDT DEBUG zjsonbin             - Serialize object to size 35b in 1ms
2020-03-20 06:20:02 EDT DEBUG zjsonbin             - Deserialized object of size 35b in 1ms
2020-03-20 06:20:02 EDT DEBUG zjsonbin             - Serialize object to size 119b in 1ms
2020-03-20 06:20:02 EDT DEBUG zjsonbin             - Deserialized object of size 119b in 1ms
2020-03-20 06:20:02 EDT DEBUG zjsonbin             - Serialize object to size 915.53Kb in 243ms
2020-03-20 06:20:02 EDT DEBUG zjsonbin             - Deserialized object of size 915.53Kb in 187ms
```

After:
```
2020-03-20 06:08:27 EDT DEBUG zjsonbin             - Serialize object to size 38b in 7ms
2020-03-20 06:08:27 EDT DEBUG zjsonbin             - Deserialized object of size 38b in 5ms
2020-03-20 06:08:27 EDT DEBUG zjsonbin             - Serialize object to size 130b in 1ms
2020-03-20 06:08:27 EDT DEBUG zjsonbin             - Deserialized object of size 130b in 0ms
2020-03-20 06:08:27 EDT DEBUG zjsonbin             - Serialize object to size 295.94Kb in 209ms
2020-03-20 06:08:27 EDT DEBUG zjsonbin             - Deserialized object of size 295.94Kb in 89ms
```

Low size files serialized better because on packing stage it removed keys with empty values that is not really affecting on overall size, but we can expect at least 3x boost on 900kb files. And looks like compressor performance is better on larger files and it would be event better after 1-2 thousands of compressions function runs, because V8 engine will optimize frequently executed functions.